Revert "move decode_and_resize to use train function arguments"
This reverts commit 49089423c2408eb69f3404182f52f68daaa38aee.
