+++ task-retrain/model.py
@@ -26,7 +26,7 @@ from sklearn.metrics import precision_score, recall_score, fbeta_score
 from utils_metrics import FBetaScore, precision_m, recall_m, fbeta_m
 from utils_readtfrecords import parse_and_augment_fn, parse_fn, get_dataset_feeder
 from utils_loss import sigmoid_focal_crossentropy
-from utils_train import zip_model_export, zip_chekpoint, model_estimator, get_optimizer
+from utils_train import zip_model_export, zip_chekpoint, model_estimator, get_optimizer, resnet_serving_input_receiver_fn
 
 
 
@@ -120,28 +120,6 @@ def train(n_classes=2, class_names=['class0', 'class1'],
     classifier = tf.estimator.add_metrics(classifier, fbeta_m)
     classifier = tf.estimator.add_metrics(classifier, precision_m)
     classifier = tf.estimator.add_metrics(classifier, recall_m)
-
-    def resnet_serving_input_receiver_fn():
-        """Convert b64 string encoded images into a tensor for production"""
-        def decode_and_resize(image_str_tensor):
-            """Decodes image string, resizes it and returns a uint8 tensor."""
-            image = tf.image.decode_image(image_str_tensor,
-                                            channels=3,
-                                            dtype=tf.uint8)
-            image = tf.reshape(image, [x_feature_shape[1], x_feature_shape[1], 3])
-            return image
-    # Run processing for batch prediction.
-        input_ph = tf.compat.v1.placeholder(tf.string, shape=[None], name='image_binary')
-        with tf.device("/cpu:0"):
-            images_tensor = tf.map_fn(decode_and_resize, input_ph, back_prop=False, dtype=tf.uint8)
-        # Cast to float
-        images_tensor = tf.cast(images_tensor, dtype=tf.float32)
-        # re-scale pixel values between 0 and 1
-        images_tensor = tf.divide(images_tensor, 255)
-
-        return tf.estimator.export.ServingInputReceiver(
-            {'input_1': images_tensor},
-            {'image_bytes': input_ph})
 
     ###############################
     # Create data feeder functions
